Update index.mdx
Line 12: Please check. If "from edge to cloud" modifies "neural networks," I would make the suggested change. If "from edge to cloud" modifies "CPU," the sentence is fine as it was.

Line 18:  I'm assuming 100% of the audience will know what FLOPS means. If not, this should be spelled out with the acronym.