Update benchmarking.mdx
Line 12: Consistency issue--CLI is spelled out here, but not in other articles.
Line 116: What is the meaning of the icon? How is it entered in Markdown?
Line 120 & 126: Is the "engine" the DeepSparse Engine? If so, it should be spelled out as such or "engine" should have an initial cap (Engine).