Update embedding-extraction.mdx
Line 54:  I'm assuming the sentence means that a local path can be passed in lieu of using a SparseZoo stub to pull down the model. If so, "also" is confusing because it implies that you can use a SparseZoo stub and also pass a local file.

Line 96:  Is it literally "two extra utilities" or for a number of extra utilities? I assumed the latter.

Line 162:  Is "spinning up" correctd?