Update logging.mdx
Line 10 and throughout:  I assume DeepSparse Logging is not a product name, so "logging" should not be capitalized.

Line 43:  This sentence is not parallel and I'm not sure of how to change it. "This facilitates" different things, but I don't think it facilitates possible data drift. Should it be "... and possible data drift detection" or something similar? The sentence has "inspection," "understanding of," and "detection." It needs something similar for possible data drift.

Line 76:  Is "dict access" an abbreviation that should be spelled out?

Line 242:  Is this the DeepSparse Server? "The DeepSparse Server startup CLI command..."
Throughout, I'm not sure when the reference is to DeepSparse Server and then just some other server. This affects capitalization.